Move key-related code to its own module
